Extract MCP external config schema package seam
