+++ tldw_Server_API/app/core/MCP_unified/tests/test_external_server_config_schema.py
@@ -2,8 +2,10 @@ from __future__ import annotations
 
 import pytest
 
+from tldw_Server_API.app.core.MCP_unified.external_servers import config_schema as config_schema_mod
 from tldw_Server_API.app.core.MCP_unified.external_servers.config_schema import (
     ExternalAuthMode,
+    ExternalServerRegistryConfig,
     parse_external_server_registry,
 )
 
@@ -108,3 +110,48 @@ def test_policy_allow_and_deny_patterns() -> None:
     assert policy.allows_tool("docs.search") is True
     assert policy.allows_tool("docs.delete") is False
     assert policy.allows_tool("ci.run") is False
+
+
+def test_standalone_package_exports_external_server_config_schema() -> None:
+    from mcp_unified import federation
+    from mcp_unified.federation import config_schema as package_schema
+
+    assert federation.ExternalAuthMode is package_schema.ExternalAuthMode
+    assert federation.ExternalMCPServerConfig is package_schema.ExternalMCPServerConfig
+    assert federation.parse_external_server_registry is package_schema.parse_external_server_registry
+    assert config_schema_mod.ExternalAuthMode is package_schema.ExternalAuthMode
+    assert config_schema_mod.ExternalServerRegistryConfig is package_schema.ExternalServerRegistryConfig
+
+    cfg = package_schema.parse_external_server_registry(
+        {
+            "servers": [
+                {
+                    "id": "docs",
+                    "name": "Docs",
+                    "transport": "websocket",
+                    "websocket": {"url": "wss://example.test/ws"},
+                }
+            ]
+        }
+    )
+
+    assert cfg.servers[0].id == "docs"
+
+
+def test_host_loader_preserves_legacy_default_config_path(monkeypatch: pytest.MonkeyPatch) -> None:
+    calls: list[tuple[str | None, str | None]] = []
+
+    def fake_package_loader(
+        config_path: str | None = None,
+        *,
+        default_config_path: str | None = None,
+    ) -> ExternalServerRegistryConfig:
+        calls.append((config_path, default_config_path))
+        return ExternalServerRegistryConfig()
+
+    monkeypatch.setattr(config_schema_mod, "_load_external_server_registry", fake_package_loader)
+
+    assert config_schema_mod.load_external_server_registry() == ExternalServerRegistryConfig()
+    assert calls == [
+        (None, "tldw_Server_API/Config_Files/mcp_external_servers.yaml"),
+    ]
